@@ -16,30 +16,34 @@
     "@react-native-community/slider": "^4.4.3",
     "@react-navigation/native": "^6",
     "@react-navigation/stack": "^6",
+    "binary-search": "^1.3.6",
+    "expo": "^49.0.0",
+    "expo-av": "~13.4.1",
     "react": "^18",
     "react-dom": "^18",
     "react-native": "^0.72",
     "react-native-fs": "^2.20.0",
     "react-native-gesture-handler": "^1.10.3",
     "react-native-safe-area-context": "^3.3.0",
-    "react-native-screens": "^3.6.0",
-    "react-native-sound": "^0.11.1"
+    "react-native-screens": "^3.6.0"
   },
   "devDependencies": {
-    "@babel/core": "7.15.0",
-    "@babel/runtime": "7.15.3",
-    "@react-native-community/eslint-config": "2.0.0",
-    "babel-jest": "26.6.3",
+    "@babel/core": "^7.20.0",
+    "@babel/preset-env": "^7.20.0",
+    "@babel/runtime": "^7.20.0",
+    "@react-native/codegen": "^0.72",
+    "@react-native/eslint-config": "^0.72.2",
+    "@react-native/metro-config": "^0.72.11",
+    "babel-jest": "^29.2.1",
     "babel-loader": "^9.1.3",
     "babel-plugin-react-native-web": "^0.19.8",
     "css-loader": "^6.8.1",
-    "eslint": "7.14.0",
+    "eslint": "^8.19.0",
     "html-bundler-webpack-plugin": "^2.11.0",
-    "jest": "26.6.3",
-    "metro-react-native-babel-preset": "^0.72",
-    "react-native-codegen": "^0.0.7",
+    "jest": "^29.2.1",
+    "metro-react-native-babel-preset": "^0.76.8",
     "react-native-web": "^0.19",
-    "react-test-renderer": "18",
+    "react-test-renderer": "18.2.0",
     "webpack": "^5.88.2",
     "webpack-cli": "^5.1.4"
   },
