@@ -35,8 +35,8 @@ const imageLoaderConfiguration = {
   type: "asset",
 };
 
-/** @type {webpack.Configuration} */
-module.exports = {
+/** @returns {webpack.Configuration} */
+module.exports = (env) => ({
 
   // configures where the build ends up
   output: {
@@ -66,7 +66,8 @@ module.exports = {
   resolve: {
     // This will only alias the exact import "react-native"
     alias: {
-      "react-native": "react-native-web",
+      "react-native$": "react-native-web",
+      "react-native/Libraries/Image/AssetRegistry": "react-native-web/dist/modules/AssetRegistry",
     },
     // If you're working on a multi-platform React Native app, web-specific
     // module implementations should be written in files using the extension
@@ -75,8 +76,9 @@ module.exports = {
   },
   plugins: [
     new webpack.EnvironmentPlugin({ REACT_NAV_LOGGING: "" }),
+    new webpack.DefinePlugin({ __DEV__: !env.production }),
     new HtmlBundlerPlugin({
       entry: { index: path.resolve(appDirectory, "web/index.html") },
     }),
   ],
-};
+});
